requirements.txt
python-telegram-bot==20.7
python-dotenv
mysql-connector-python
google-generativeai
